Major cleanup and reorganization of Nymo Art v4 project
🧹 MASSIVE PROJECT CLEANUP:
- Removed 25+ unused/empty files from project root
- Deleted redundant 'my_api/' directory entirely
- Cleaned up empty Python files and old backup files
- Removed duplicate PhotoReal and frontend backup files

📁 NEW DIRECTORY STRUCTURE:
- Created 'scripts/' directory for CLI tools
- Moved generate_test_images.py and batch_process.py to scripts/
- Fixed import paths for relocated CLI scripts
- Updated backend_path references in CLI tools

🔧 BACKEND IMPROVEMENTS:
- Added new Leonardo AI engines: Flux and PhotoReal
- Enhanced Phoenix engine with proper configuration
- Fixed class name: LeonardoPhoenixEngine → PhoenixEngine
- Added comprehensive API routes for all engines
- Improved batch processing capabilities
- Updated unit tests and validation

🎨 FRONTEND COMPLETE REWRITE:
- Built modern React TypeScript frontend with Vite
- Added Tailwind CSS for beautiful, responsive design
- Created modular components for each AI engine
- Implemented proper form handling and validation
- Added image gallery and download functionality
- Multi-engine support with dedicated pages

📋 CLI TOOLS ENHANCED:
- Fixed style name mappings for Leonardo AI
- Updated to use correct engine configurations
- Added comprehensive error handling
- Improved batch processing workflow

📚 DOCUMENTATION & ORGANIZATION:
- Created comprehensive PROJECT_STRUCTURE.md
- Enhanced .gitignore with proper exclusions
- Organized all files into logical directories
- Removed development artifacts and test images

This commit transforms the project from a chaotic collection of files
into a clean, modular, production-ready architecture with working
CLI tools, comprehensive backend API, and modern frontend interface.

+++ backend/core/engine/base.py
@@ -86,7 +86,7 @@ class TextGenerationEngine(BaseEngine):
     
     def estimate_tokens(self, text: str) -> int:
         """Estimate token count. Simple approximation, override for accuracy."""
-        return len(text.split()) * 1.3  # Rough approximation
+        return int(len(text.split()) * 1.3)  # Rough approximation
 
 
 class EngineRegistry:
